Move removal of old deleted files to a new function
Since removing of all of the matching deleted files between a build with
an older manifest format and current needs to happen *after*
match_manifests() finishes accounting for new and deleted entries in the
manifests, simply move the logic into a separate function to be called
after match_manifests().

Because one or more file deletions should trigger a fresh manifest
creation, account for the return value of this new function in addition
to the result of manifest pruning and whether any bundle includes were
changed.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>