swupd_create_pack: download original files on demand for diffing
The new, optional mode gets enabled via the --content-url parameter.
When specified, swupd_create_pack will download the <original
version>/files/<hash>.tar file from the update server if the required
file is missing.

This mode is meant for use with meta-swupd where new CI builds start
with an empty "image" directory. NFS access to the full, expanded
"image" directories of the previous build is not possible in such a
setup (would require root for special file attributes).

Downloading and unpacking image directories of previous builds would
be possible, but is expected to be slower (when downloading all files,
not just those needed for diffing) or more complex (when deciding about
required files outside of swupd_create_pack).

The downside of this new approach is that it relies on not changing
the content or naming of the tar files. For example, switching from
GNU tar to bsdtar would not work. But that's not an issue for the
intended usage in Ostro OS, which already made the switch to
bsdtar.

If for some reason a format change is necessary, delta computation
becomes impossible and falls back to staging the full files.

Upstream-Status: Submitted [https://github.com/clearlinux/swupd-server/pull/47]

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>

%% original patch: 0003-swupd_create_pack-download-original-files-on-demand-.patch