type_change.c: allow transition dir -> symlink
swupd client (tested with 2.87) supports replacing an empty directory
with a symlink, so the check in the server is overly restrictive.

Upstream wants to invest some more work before officially supporting
such a state change (https://github.com/clearlinux/swupd-server/issues/50).

In the meantime we already need to support it for updates from Ostro
OS 1.0 because of /usr/lib/ssl/certs changing to a symlink.

Upstream-Status: Inappropriate [Workaround]

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>

%% original patch: 0001-type_change.c-allow-transition-dir-symlink.patch